requirements.txt
streamlit
psycopg2-binary
pandas
altair
Image
fetch_dashboard_data